backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
scikit-learn==1.3.2
Pillow==10.2.0
werkzeug==3.0.1
datetime
bson